+++ providers/google/src/airflow/providers/google/cloud/hooks/cloud_memorystore.py
@@ -45,7 +45,7 @@ from google.cloud.redis_v1 import (
 )
 
 from airflow import version
-from airflow.exceptions import AirflowException
+from airflow.providers.common.compat.sdk import AirflowException
 from airflow.providers.google.common.hooks.base_google import PROVIDE_PROJECT_ID, GoogleBaseHook
 
 if TYPE_CHECKING: