+++ task-sdk/src/airflow/sdk/serde/__init__.py
@@ -30,9 +30,9 @@ from typing import TYPE_CHECKING, Any, TypeVar, cast
 import attr
 
 from airflow.observability.stats import Stats
+from airflow.sdk._shared.module_loading import import_string, iter_namespace, qualname
 from airflow.sdk.configuration import conf
 from airflow.sdk.serde.typing import is_pydantic_model
-from airflow.utils.module_loading import import_string, iter_namespace, qualname
 
 if TYPE_CHECKING:
     from types import ModuleType