+++ task-sdk/src/airflow/sdk/serde/serializers/builtin.py
@@ -19,7 +19,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, cast
 
-from airflow.utils.module_loading import qualname
+from airflow.sdk.module_loading import qualname
 
 if TYPE_CHECKING:
     from airflow.sdk.serde import U