+++ task-sdk/src/airflow/sdk/serde/serializers/deltalake.py
@@ -19,7 +19,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from airflow.utils.module_loading import qualname
+from airflow.sdk.module_loading import qualname
 
 serializers = ["deltalake.table.DeltaTable"]
 deserializers = serializers